+++ documents/license.mdx
@@ -1,16 +1,12 @@
 ---
-title: Hello
+title: はじめに
 slug: home
 ---
-
-Modified BSD License
-====================
-
 _Copyright © {props.year}, {props.holder}
-_All rights reserved._
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
+
 
 (省略)
