+++ src/pages/index.tsx
@@ -1,24 +1,37 @@
-import { Inter } from 'next/font/google'
-import Head from 'next/head'
-import Image from 'next/image'
+import { GetStaticProps } from 'next'
+import Link from 'next/link'
 import { Button } from 'smarthr-ui'
 
-import styles from '@/styles/Home.module.css'
+import { PostData, getSortedPostsData } from '../libs/mdxUtils'
 
-export default function Home() {
+interface ArchiveProps {
+  allPostsData: PostData[]
+}
+
+export default function Archive({ allPostsData }: ArchiveProps) {
   return (
-    <>
-      <Head>
-        <title>Create Next App</title>
-        <meta name="description" content="Generated by create next app" />
-        <meta name="viewport" content="width=device-width, initial-scale=1" />
-        <link rel="icon" href="/favicon.ico" />
-      </Head>
-      <main className={styles.main}>
-        <p>
-          <Button>Text</Button>
-        </p>
-      </main>
-    </>
+    <main>
+      <h1>LEARN SmartHR QUICK START</h1>
+      <ul>
+        {allPostsData.map(({ id, date, title }) => (
+          <li key={id}>
+            <Link href={`/documents/${id}`} passHref>
+              {title}
+            </Link>
+            <br />
+            {date}
+          </li>
+        ))}
+      </ul>
+    </main>
   )
+}
+
+export const getStaticProps: GetStaticProps<ArchiveProps> = async () => {
+  const allPostsData = getSortedPostsData()
+  return {
+    props: {
+      allPostsData,
+    },
+  }
 }
